Invert's extruder 0 direction
We had to do this in order to have the direct drive work, one of the motors needed to be inversed. 

This did fix the direct drive but we need to see if it will breack the bowden setup if so this needs to be applied for the other extruder and undoned for this one.